fix sign-compare between long unsigned int and int

+++ src/game/material.cpp
@@ -48,7 +48,7 @@ bool CMatDefault::Get(const char *pName, float *pValue) const
 
 CMatDefault &CMaterials::operator[](int Index) const
 {
-	if(Index >= ms_aMaterials.size())
+	if(Index >= (int)ms_aMaterials.size())
 		return const_cast<CMatDefault &>(ms_aMaterials[0]); //Unknown material, maybe the map is too new?
 	return const_cast<CMatDefault &>(ms_aMaterials[Index]);
 }
